Replaced invocations of 'canonicalizePath' with calls to a new function, 'ensureAbsolute', in a newly added Utilities module.
This avoids issues with the new behavior of canonicalizePath in recent versions of Directory, namely that it throws exceptions when the path does not exist.  ensureAbsolute is the identity function for absolute paths (to the degree possible, it is in IO), and roots relative paths in the current directory.

These changes should be sufficient to use cabal-dev with ghc 7.